fix up static html pages
