+++ src/landing/logo.jsx
@@ -1,11 +1,7 @@
 // @ts-check
 
 import React, { useState, useEffect, useCallback } from 'react';
-
-import logoDay from '../assets/CleardayLarge.png';
-import logoNight from '../assets/ClearnightLarge.png';
-import logoDaySmall from '../assets/CleardaySmall.png';
-import logoNightSmall from '../assets/ClearnightSmall.png';
+import { getLogoImage } from '../logo-images';
 
 function useForceUpdate() {
   const [, update] = useState(0);
@@ -33,48 +29,15 @@ export function Logo({ className, ...rest }) {
     };
   }, [images.large]);
 
-  const img = <img srcSet="" />;
   return (
     <figure className={'home-header-logo ' + (className || '')}>
       <img
         {...rest}
         alt="clearsky logo"
         src={images.large}
-        srcSet={`${images.small} 800w, ${images.large} 1700w`}
+        srcSet={images.srcSet}
         sizes="(max-width: 800px) 50vw, 100vw"
       />
     </figure>
   );
-}
-
-function getLogoImage() {
-  const currentTime = new Date();
-  const hours = currentTime.getHours();
-
-  const DAY_START = 6;
-  const DAY_END = 18;
-
-  let useDay = false;
-  const refreshIn = new Date(currentTime);
-  refreshIn.setMinutes(0);
-  refreshIn.setSeconds(0);
-  refreshIn.setMilliseconds(0);
-
-  if (hours < DAY_START) {
-    refreshIn.setHours(DAY_START);
-  } else if (hours > DAY_END) {
-    refreshIn.setHours(DAY_START);
-    refreshIn.setDate(refreshIn.getDate() + 1);
-  } else {
-    useDay = true;
-    refreshIn.setHours(DAY_END);
-  }
-
-  const large = useDay ? logoDay : logoNight;
-  const small = useDay ? logoDaySmall : logoNightSmall;
-  return {
-    large,
-    small,
-    refreshInMsec: refreshIn.getTime() - currentTime.getTime(),
-  };
 }
