+++ tsconfig.json
@@ -11,5 +11,5 @@
     "skipLibCheck": true,
     "noEmit": true
   },
-  "files": ["src/app.jsx", "src/types.d.ts"]
+  "files": ["src/app.jsx", "src/logo-images.js", "src/types.d.ts"]
 }
